Extract shared PeriodPicker for metrics views
