+++ Sources/Scout/UI/Activity/ActivityView.swift
@@ -20,7 +20,7 @@ struct ActivityView: View {
 
     var body: some View {
         VStack(spacing: 0) {
-            PeriodPicker(extent: $extent, periods: ActivityPeriod.allCases)
+            PeriodExtentPicker(extent: $extent, periods: ActivityPeriod.allCases)
 
             ProviderView(provider: activity) { data in
                 RangeControl(extent: $extent)
