+++ Sources/Scout/UI/Metrics/MetricsView.swift
@@ -22,15 +22,7 @@ struct MetricsView<T: ChartNumeric>: View {
     }
 
     var body: some View {
-        Picker(selection: $extent.period) {
-            ForEach(Period.allCases) { period in
-                Text(period.shortTitle.uppercased())
-            }
-        } label: {
-            Text(verbatim: "Period")
-        }
-        .padding(.horizontal)
-        .pickerStyle(.segmented)
+        PeriodPicker(selection: $extent.period)
 
         RangeControl(extent: $extent)
