+++ Sources/Scout/UI/Stats/StatView.swift
@@ -19,7 +19,7 @@ struct StatView: View {
 
     var body: some View {
         VStack(spacing: 0) {
-            PeriodPicker(extent: $extent, periods: stat.periods)
+            PeriodExtentPicker(extent: $extent, periods: stat.periods)
 
             ProviderView(provider: stat) { data in
                 RangeControl(extent: $extent)
